\name{dg}
\alias{dg}

\title{
 get nodes by regree
}

\description{
 get nodes by regree
}

\usage{
  dg(g, gene, degree=1);
}

\arguments{
   \item{g}{igraph object}
   \item{gene}{character: gene}
   \item{degree}{positive integer: degree}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

