\name{diffusion}
\alias{diffusion}

\title{
 calculate diffusion score
}

\description{
 calculate diffusion score
}

\usage{
  diffusion(g, gene, cutoff);
}

\arguments{
   \item{g}{igraph object}
   \item{gene}{character: gene}
   \item{cutoff}{numeric: cutoff}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

