\name{find_graph_with_genes}
\alias{find_graph_with_genes}

\title{
 find graph with genes
}

\description{
 find graph with genes
}

\usage{
  find_graph_with_genes(gene_set, graph_list);
}

\arguments{
   \item{gene_set}{character: gene}
   \item{graph_list}{list: graphs}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

