\name{ncluster}
\alias{ncluster}

\title{
 do cluster for network propagation
}

\description{
 do cluster for network propagation
}

\usage{
  ncluster(g, gene, cutoff);
}

\arguments{
   \item{g}{igraph object}
   \item{gene}{character: gene}
   \item{cutoff}{the maximum number of genes contained in networks}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

