%% main
\name{NetProp-package}
\alias{NetProp}
\title{
 NetProp package
}
\description{
 NetProp tools.
}
\author{
%%  ~~ possibly secondary sources and usages ~~
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}
\references{
%%  ~~ possibly secondary sources and usages ~~
 NetProp: an R package for network propagation
}
%%\keyword{key}
