\name{rwr}
\alias{rwr}

\title{
 do random walk with restart 
}

\description{
 do random walk with restart 
}

\usage{
  rwr(g, gene, cutoff);
}

\arguments{
   \item{g}{igraph object}
   \item{gene}{character: gene}
   \item{cutoff}{numeric: cutoff}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

