\name{sp}
\alias{sp}

\title{
 do the shortest path
}

\description{
 do the shortest path
}

\usage{
  sp(g, gene);
}

\arguments{
   \item{g}{igraph object}
   \item{gene}{character: gene}
}

\author{
 Ying Hu <yhu@mail.nih.gov>
 Chunhua Yan <yanch@mail.nih.gov>
}

\references{
##
}

\examples{
##
}

